requirements.txt
cmake
dlib
face_recognition
opencv-python
pillow